Updated rules command anyone can  recall a rule

+++ src/commands/Information/Server Info/rules.ts
@@ -1,5 +1,4 @@
-import { SteveCommand } from '@lib/structures/commands/SteveCommand';
-import { PermissionLevels } from '@lib/types/enums';
+mport { SteveCommand } from '@lib/structures/commands/SteveCommand';
 import { CommandStore, KlasaMessage } from 'klasa';
 import { Message } from 'discord.js';
 
@@ -8,24 +7,40 @@ export default class extends SteveCommand {
 	public constructor(store: CommandStore, file: string[], directory: string) {
 		super(store, file, directory, {
 			description: 'Display the rules',
-			permissionLevel: PermissionLevels.ADMINISTRATOR
+			aliases: ['rule'],
+			runIn: ['text'],
+			subcommands: true,
+			usage: '<all|view:default> (number:number)'
 		});
+		this
+			.createCustomResolver('number', (num, possible, msg, [action]) => action !== 'all' ? num : null);
 	}
 
-	async run(msg: KlasaMessage): Promise<Message> {
-		const rules: string[] = [
-			'[0] Be respectful, there are no dumb questions.',
-			'[1] Your nickname should be your real name, especially if you want participation points!',
-			'[2] Don\'t spam. Don\'t spam emojis, pictures, messages, pings, etc. Just don\'t spam.',
-			'[3] Please keep to relevant channel topics, and if a Staff member or Admin asks you to move your conversation, listen.',
-			'[4] We encourage you to help each other! However, saying  things like "It works on my machine" or "I don\'t have the problem" isn\'t helpful.',
-			'[5] Keep your language clean- don\'t say something you wouldn\'t say to your professor.',
-			'[6] Make sure to read the [#useful-info] channel! As the name implies, there is useful info in there.',
-			'[7] Respect what the Staff and Admins say, they are in charge for a reason. If they tell you to stop, then stop.'
-		];
+	rules: string[] = [
+		'Be respectful, there are no dumb questions.',
+		'Your nickname should be your real name, especially if you want participation points!',
+		'Don\'t spam. Don\'t spam emojis, pictures, messages, pings, etc. Just don\'t spam.',
+		'Please keep to relevant channel topics, and if a Staff member or Admin asks you to move your conversation, listen.',
+		'We encourage you to help each other! However, saying  things like "It works on my machine" or "I don\'t have the problem" isn\'t helpful.',
+		'Keep your language clean- don\'t say something you wouldn\'t say to your professor.',
+		'Make sure to read the [#useful-info] channel! As the name implies, there is useful info in there.',
+		'Respect what the Staff and Admins say, they are in charge for a reason. If they tell you to stop, then stop.'
+	];
+
+	async all(msg: KlasaMessage): Promise<Message> {
+		if (!msg.member.isAdmin) throw 'Sorry but you cant do that';
 		msg.delete();
+		let buff = '';
+		this.rules.forEach((rule, idx) => {
+			buff += `[${idx}] ${rule}\n`;
+		});
 		await msg.channel.send({ files: [{ attachment: './assets/images/rules.png', name: 'rules.png' }] });
-		return msg.sendCode('ini', rules.join('\n'));
+		return msg.sendCode('ini', buff);
+	}
+
+	async view(msg: KlasaMessage, ruleNum: number): Promise<Message> {
+		if (ruleNum >= this.rules.length) throw `There is no rule ${ruleNum}`;
+		return msg.channel.send(this.rules[ruleNum]);
 	}
 
 }
